Asks for configuration data if it wasn't found or the located data was corrupt.
Another step towards v0.1.3! In addition to the aforementioned feature, a few changes have been made. Most notably, the googleConfig.json file now refers to the spreadsheet as *SheetsID* rather than *SpreadsheetID*. Progress towards 0.1.3 is smoothly continuing, however, I'm not completely sure if I will be able to complete it for tomorrow. If not on Sunday, then the update will almost certainly be released the next day.

*Summary*
This commit resolves two issues, #7 and #6 .