PM Users after successful role/nickname application
Checks to see if any new roles are added or the nickname is changed. If they aren't, then the user is not PMed. It is possible to recieve multiple messages if a role is removed or the user's nickname is changed without changing it in the sheet.

Some performance optimizations included